'140' Fix alias code to stop deletion of dsapp in folder

+++ dsapp-test.sh
@@ -13,7 +13,7 @@
 #	Declaration of Variables
 #
 ##################################################################################################
-	dsappversion='139'
+	dsappversion='140'
 	autoUpdate=true
 	dsappDirectory="/opt/novell/datasync/tools/dsapp"
 	dsappLogs="$dsappDirectory/logs"
@@ -202,32 +202,36 @@ function setupDsappAlias {
 			resetEnvironment=true
 		fi
 		
-		# Check if running version is newer than installed version
-		installedVersion=`grep -m1 dsappversion= /opt/novell/datasync/tools/dsapp/dsapp.sh 2>/dev/null | cut -f2 -d "'"`
-		if [[ "$dsappversion" -gt "$installedVersion" ]];then
-			tellUserAboutAlias=true
-			echo "Installing dsapp to /opt/novell/datasync/tools/dsapp/"
-			mv -v dsapp.sh $dsappDirectory
-			if [ $? -ne 0 ]; then
-				echo -e "\nThere was a problem copying dsapp.sh to /opt/novell/datasync/tools/dsapp..."
+		#Skip if already in dsappDirectory
+        if [[ "$PWD" != "$dsappDirectory" ]];then
+
+			# Check if running version is newer than installed version
+			installedVersion=`grep -m1 dsappversion= /opt/novell/datasync/tools/dsapp/dsapp.sh 2>/dev/null | cut -f2 -d "'"`
+			if [[ "$dsappversion" -gt "$installedVersion" ]];then
+				tellUserAboutAlias=true
+				echo "Installing dsapp to /opt/novell/datasync/tools/dsapp/"
+				mv -v dsapp.sh $dsappDirectory
+				if [ $? -ne 0 ]; then
+					echo -e "\nThere was a problem copying dsapp.sh to /opt/novell/datasync/tools/dsapp..."
+				fi
+			else 
+				tellUserAboutAlias=true
+				rm dsapp.sh
 			fi
-		else 
-			tellUserAboutAlias=true
-			rm dsapp.sh
-		fi
 
-		# Reset environment variables (loads /etc/profile for dsapp alias)
-		if ($resetEnvironment); then
-			echo -e "Refreshing environment variables..."
-			su -
-		fi
+			# Reset environment variables (loads /etc/profile for dsapp alias)
+			if ($resetEnvironment); then
+				echo -e "Refreshing environment variables..."
+				su -
+			fi
 
-		if($tellUserAboutAlias); then
-			echo -e "\nPlease use /opt/novell/datasync/tools/dsapp/dsapp.sh"
-			echo -e "To launch, enter the following anywhere: dsapp\n"
-		fi
+			if($tellUserAboutAlias); then
+				echo -e "\nPlease use /opt/novell/datasync/tools/dsapp/dsapp.sh"
+				echo -e "To launch, enter the following anywhere: dsapp\n"
+			fi
 
-		exit 0
+			exit 0
+		fi
 	fi
 }
